Add custom wrapper to lalrpop so that errors are 'static
By default, lalrpop used &str as part of the result type, and these had
the same lifetime as the input to the parser. However, this caused
problems with the failure crate, which expects all errors to have teh
static lifetime. To fix this, a simple wrapper was added which maps the
&str to Strings.